fix: alert 위치 이동 - 여러 개 렌더링되는 BookInfoWrapper에 너무 많은 기능이 가지 않도록
